Issue #7002: Extract icon decoder into a generic image decoder in support-images.
Extracts `AndroidIconDecoder`, `IconDecoder` and `DesiredSize` out of `browser-icons`
into a new component `support-images`, which provides helpers for handling images. `AndroidIconDecoder` and `IconDecoder`
are renamed to `AndroidImageDecoder` and `ImageDecoder` in `support-images`.
