+++ components/browser/icons/build.gradle
@@ -40,10 +40,11 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
 }
 
 dependencies {
-    implementation project(':support-ktx')
     implementation project(':concept-engine')
     implementation project(':concept-fetch')
     implementation project(':browser-state')
+    implementation project(':support-images')
+    implementation project(':support-ktx')
 
     implementation Dependencies.androidx_annotation
     implementation Dependencies.androidx_core_ktx
