+++ components/browser/icons/src/main/java/mozilla/components/browser/icons/decoder/ICOIconDecoder.kt
@@ -5,9 +5,10 @@
 package mozilla.components.browser.icons.decoder
 
 import android.graphics.Bitmap
-import mozilla.components.browser.icons.DesiredSize
+import mozilla.components.support.images.DesiredSize
 import mozilla.components.browser.icons.decoder.ico.decodeDirectoryEntries
 import mozilla.components.browser.icons.utils.findBestSize
+import mozilla.components.support.images.decoder.ImageDecoder
 
 // Some geometry of an ICO file.
 internal const val HEADER_LENGTH_BYTES = 6
@@ -16,12 +17,12 @@ internal const val ICON_DIRECTORY_ENTRY_LENGTH_BYTES = 16
 internal const val ZERO_BYTE = 0.toByte()
 
 /**
- * [IconDecoder] implementation for decoding ICO files.
+ * [ImageDecoder] implementation for decoding ICO files.
  *
  * An ICO file is a container format that may hold up to 255 images in either BMP or PNG format.
  * A mixture of image types may not exist.
  */
-class ICOIconDecoder : IconDecoder {
+class ICOIconDecoder : ImageDecoder {
     override fun decode(data: ByteArray, desiredSize: DesiredSize): Bitmap? {
         val (targetSize, maxSize, maxScaleFactor) = desiredSize
         val entries = decodeDirectoryEntries(data, maxSize)
