+++ components/browser/icons/src/main/java/mozilla/components/browser/icons/processor/DiskIconProcessor.kt
@@ -5,10 +5,10 @@
 package mozilla.components.browser.icons.processor
 
 import android.content.Context
-import mozilla.components.browser.icons.DesiredSize
 import mozilla.components.browser.icons.Icon
 import mozilla.components.browser.icons.Icon.Source
 import mozilla.components.browser.icons.IconRequest
+import mozilla.components.support.images.DesiredSize
 
 /**
  * [IconProcessor] implementation that saves icons in the disk cache.
