+++ components/browser/icons/src/test/java/mozilla/components/browser/icons/decoder/ICOIconDecoderTest.kt
@@ -5,9 +5,9 @@
 package mozilla.components.browser.icons.decoder
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import mozilla.components.browser.icons.DesiredSize
 import mozilla.components.browser.icons.decoder.ico.IconDirectoryEntry
 import mozilla.components.browser.icons.decoder.ico.decodeDirectoryEntries
+import mozilla.components.support.images.DesiredSize
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Test
