+++ components/support/images/proguard-rules-consumer.pro
@@ -1,0 +1,1 @@
+# ProGuard rules for consumers of this library.
